Consolidate STT language remapping into metadata command
Removed the separate remap_stt_language command and its tests, moving language remapping logic into remap_stt_metadata. Updated remap_stt_metadata and its tests to handle language field updates for items and agenda resources, ensuring consistent processing and reducing code duplication.

STT-1223
