+++ server/stt/commands/__init__.py
@@ -1,2 +1,1 @@
-from . import remap_stt_language  # noqa
 from . import remap_stt_metadata  # noqa
